refactor(native): modularize design token constants
