+++ apps/native/lib/constants/theme.ts
@@ -1,0 +1,80 @@
+import { DARK_ELEVATION_TOKENS, LIGHT_ELEVATION_TOKENS, SHARED_TOKENS } from "./foundations";
+
+export const NAV_THEME = {
+  light: {
+    background: "hsl(0, 0%, 100%)",
+    foreground: "hsl(235.5556, 14.7541%, 35.8824%)",
+    text: "hsl(235.5556, 14.7541%, 35.8824%)",
+    card: "hsl(0, 0%, 100%)",
+    cardForeground: "hsl(235.5556, 14.7541%, 35.8824%)",
+    popover: "hsl(222.0000, 11.3636%, 82.7451%)",
+    popoverForeground: "hsl(235.5556, 14.7541%, 35.8824%)",
+    primary: "hsl(266.0440, 85.0467%, 58.0392%)",
+    primaryForeground: "hsl(0, 0%, 100%)",
+    secondary: "hsl(222.0000, 11.3636%, 82.7451%)",
+    secondaryForeground: "hsl(235.5556, 14.7541%, 35.8824%)",
+    muted: "hsl(222.0000, 17.8571%, 89.0196%)",
+    mutedForeground: "hsl(231.8182, 9.1667%, 47.0588%)",
+    accent: "hsl(270, 59.2593%, 57.6471%)",
+    accentForeground: "hsl(0, 0%, 100%)",
+    destructive: "hsl(347.0769, 86.6667%, 44.1176%)",
+    destructiveForeground: "hsl(0, 0%, 100%)",
+    border: "hsl(223.3333, 15.5172%, 77.2549%)",
+    input: "hsl(222.0000, 11.3636%, 82.7451%)",
+    ring: "hsl(266.0440, 85.0467%, 58.0392%)",
+    notification: "hsl(347.0769, 86.6667%, 44.1176%)",
+    chart1: "hsl(17.8776, 100%, 48.0392%)",
+    chart2: "hsl(174.8000, 100%, 29.4118%)",
+    chart3: "hsl(195.7143, 72.4138%, 22.7451%)",
+    chart4: "hsl(43.5294, 100%, 50%)",
+    chart5: "hsl(36.3780, 100%, 49.8039%)",
+    sidebar: "hsl(218.1818, 25.5814%, 91.5686%)",
+    sidebarForeground: "hsl(235.5556, 14.7541%, 35.8824%)",
+    sidebarPrimary: "hsl(266.0440, 85.0467%, 58.0392%)",
+    sidebarPrimaryForeground: "hsl(0, 0%, 100%)",
+    sidebarAccent: "hsl(270, 59.2593%, 57.6471%)",
+    sidebarAccentForeground: "hsl(0, 0%, 100%)",
+    sidebarBorder: "hsl(223.3333, 15.5172%, 77.2549%)",
+    sidebarRing: "hsl(266.0440, 85.0467%, 58.0392%)",
+    ...SHARED_TOKENS,
+    ...LIGHT_ELEVATION_TOKENS,
+  },
+  dark: {
+    background: "hsl(240, 22.7273%, 8.6275%)",
+    foreground: "hsl(226.1538, 63.9344%, 88.0392%)",
+    text: "hsl(226.1538, 63.9344%, 88.0392%)",
+    card: "hsl(240, 22.7273%, 8.6275%)",
+    cardForeground: "hsl(226.1538, 63.9344%, 88.0392%)",
+    popover: "hsl(240, 21.3115%, 11.9608%)",
+    popoverForeground: "hsl(226.1538, 63.9344%, 88.0392%)",
+    primary: "hsl(267.4074, 83.5052%, 80.9804%)",
+    primaryForeground: "hsl(240, 22.7273%, 8.6275%)",
+    secondary: "hsl(236.8421, 16.2393%, 22.9412%)",
+    secondaryForeground: "hsl(226.1538, 63.9344%, 88.0392%)",
+    muted: "hsl(236.8421, 16.2393%, 22.9412%)",
+    mutedForeground: "hsl(227.6471, 23.6111%, 71.7647%)",
+    accent: "hsl(231.8919, 97.3684%, 85.0980%)",
+    accentForeground: "hsl(240, 22.7273%, 8.6275%)",
+    destructive: "hsl(343.2692, 81.2500%, 74.9020%)",
+    destructiveForeground: "hsl(240, 22.7273%, 8.6275%)",
+    border: "hsl(234.2857, 13.2075%, 31.1765%)",
+    input: "hsl(236.8421, 16.2393%, 22.9412%)",
+    ring: "hsl(267.4074, 83.5052%, 80.9804%)",
+    notification: "hsl(343.2692, 81.2500%, 74.9020%)",
+    chart1: "hsl(225.4286, 84.0000%, 49.0196%)",
+    chart2: "hsl(159.8936, 100%, 36.8627%)",
+    chart3: "hsl(36.3780, 100%, 49.8039%)",
+    chart4: "hsl(273.4054, 100.0000%, 63.7255%)",
+    chart5: "hsl(345.4709, 100%, 56.2745%)",
+    sidebar: "hsl(240, 22.7273%, 8.6275%)",
+    sidebarForeground: "hsl(226.1538, 63.9344%, 88.0392%)",
+    sidebarPrimary: "hsl(267.4074, 83.5052%, 80.9804%)",
+    sidebarPrimaryForeground: "hsl(240, 22.7273%, 8.6275%)",
+    sidebarAccent: "hsl(231.8919, 97.3684%, 85.0980%)",
+    sidebarAccentForeground: "hsl(240, 22.7273%, 8.6275%)",
+    sidebarBorder: "hsl(234.2857, 13.2075%, 31.1765%)",
+    sidebarRing: "hsl(267.4074, 83.5052%, 80.9804%)",
+    ...SHARED_TOKENS,
+    ...DARK_ELEVATION_TOKENS,
+  },
+} as const;
